dx: require dependency array for useDrag and useDrop hooks
There is a subtle difference between the dependency arrays of useDrag/useDrop and
useEffect/useMemo: When you don't pass a dependeny array to useDrag/useDrop, it
uses an empty array, which will cause it to never re-execute, instead of always
re-executing like useEffect/useMemo would without a dependency array.

To make sure we never accidentally forget the dependency array again, this commit
adds anew lint rule that triggers an error if only one argument is provided to
useDrag/useDrop.

// cc @corinnaj