fix: Make dice templates work without a main character set, closes #192
This was caused because the call to `useMyActiveCharacter` can return
`null`, even though a character is selected on the map. This is because
the calculation of which characters are selected happened as part of a
chain of `useEffect`s, which cause multiple intermittent re-renders,
until the main character is properly calculated and returned.

This commit both makes the `DiceTemplate` component more robust by
explicitly checking for `null`, and also replaces one `useEffect` by
`useMemo` to shorten the chain of `useEffect`s.